move booklistfragment to store module and change class name
