+++ Features/Bookmark/build.gradle
@@ -30,6 +30,7 @@ android {
 }
 
 dependencies {
+    implementation project(':Presentation')
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
