+++ Features/Store/src/main/java/com/gymix/store/StoreFragment.kt
@@ -1,4 +1,4 @@
-package com.gymix.presentation.book
+package com.gymix.store
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -13,6 +13,7 @@ import androidx.lifecycle.repeatOnLifecycle
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.gymix.common.utils.GridManager
 import com.gymix.common.utils.network.RemoteStatus
 import com.gymix.presentation.databinding.FragmentBookListBinding
 import dagger.hilt.android.AndroidEntryPoint
@@ -20,7 +21,7 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
-class BookListFragment : Fragment() {
+class StoreFragment : Fragment() {
 
     private val viewModel: BookListViewModel by viewModels()
     private lateinit var binding: FragmentBookListBinding
