+++ app/build.gradle
@@ -52,6 +52,8 @@ dependencies {
     implementation project(path: ":Domain")
     implementation project(path: ":Presentation")
     implementation project(path: ":Common")
+    implementation project(':Features:Store')
+    implementation project(':Features:Bookmark')
 
     implementation 'androidx.multidex:multidex:2.0.1'
